[UMPA-256] Refactor : 공통 상수 분리
