+++ Umpa/Presentation/Main/TeacherFinder/ServiceDetail/Common/ServiceDetailConstant.swift
@@ -1,0 +1,8 @@
+// Created for Umpa in 2025
+
+import Foundation
+
+enum ServiceDetailConstant {
+  static let bottomActionBarHeight: CGFloat = fs(64)
+  static let thumbnailHeight: CGFloat = fs(300)
+}
