+++ Umpa/Presentation/Main/TeacherFinder/ServiceDetail/LessonService/LessonServiceDetailView.swift
@@ -5,106 +5,108 @@ import Domain
 import Factory
 import SwiftUI
 
-struct LessonServiceDetailView: ServiceDetailView {
-    @InjectedObject(\.appState) private var appState
+struct LessonServiceDetailView: View {
+  @InjectedObject(\.appState) private var appState
 
-    #if DEBUG
+  #if DEBUG
     @Injected(\.stubServiceDetailInteractor) private var serviceDetailInteractor
-    #else
+  #else
     @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
-    #endif
+  #endif
 
-    let service: LessonService
+  let service: LessonService
 
-    var tabItems: [TabItem] {
-        service.curriculum.isEmpty
-            ? [.teacherOverview, .lessonOverview, .review]
-            : [.teacherOverview, .lessonOverview, .curriculum, .review]
-    }
+  var tabItems: [TabItem] {
+    service.curriculum.isEmpty
+      ? [.teacherOverview, .lessonOverview, .review]
+      : [.teacherOverview, .lessonOverview, .curriculum, .review]
+  }
 
-    @State private var tabSelection = 0
+  @State private var tabSelection = 0
 
-    var body: some View {
-        content
-            .modifier(NavigationBackButton(.arrowBack))
-            .navigationDestination(for: ChatRoom.self) { chatRoom in
-                ChatRoomView(chatRoom: chatRoom)
-            }
-    }
+  // MARK: View
 
-    @ViewBuilder
-    var content: some View {
-        ZStack(alignment: .bottom) {
-            ScrollView {
-                VStack(spacing: fs(0)) {
-                    Header(tabSelection: $tabSelection, service: service)
-                    segmentedControlContent
-                        .frame(maxWidth: .infinity, maxHeight: .infinity)
-                }
-                .padding(.bottom, bottomActionBarHeight)
-            }
+  var body: some View {
+    content
+      .modifier(NavigationBackButton(.arrowBack))
+      .navigationDestination(for: ChatRoom.self) { chatRoom in
+        ChatRoomView(chatRoom: chatRoom)
+      }
+  }
 
-            BottomActionBar(
-                height: bottomActionBarHeight,
-                isLiked: false, // TODO: isLiked 를 받아와야 함
-                likeButtonAction: { isLiked in
-                    serviceDetailInteractor.markAsLike(isLiked, for: service.id)
-                },
-                primaryButtonAction: {
-                    serviceDetailInteractor.startChat(
-                        with: service.eraseToAnyService(),
-                        navigationPath: $appState.routing.teacherFinderNavigationPath
-                    )
-                }
-            )
+  @ViewBuilder
+  var content: some View {
+    ZStack(alignment: .bottom) {
+      ScrollView {
+        VStack(spacing: fs(0)) {
+          Header(tabSelection: $tabSelection, service: service)
+          segmentedControlContent
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
         }
-    }
+        .padding(.bottom, ServiceDetailConstant.bottomActionBarHeight)
+      }
 
-    @ViewBuilder
-    var segmentedControlContent: some View {
-        switch tabItems[tabSelection] {
-        case .teacherOverview:
-            TeacherOverviewTabContent(teacher: service.author)
-        case .lessonOverview:
-            LessonOverviewTabContent(service: service)
-        case .curriculum:
-            CurriculumTabContent(curriculumList: service.curriculum)
-        case .review:
-            ReviewTabContent(service: service.eraseToAnyService())
+      BottomActionBar(
+        height: ServiceDetailConstant.bottomActionBarHeight,
+        isLiked: false, // TODO: isLiked 를 받아와야 함
+        likeButtonAction: { isLiked in
+          serviceDetailInteractor.markAsLike(isLiked, for: service.id)
+        },
+        primaryButtonAction: {
+          serviceDetailInteractor.startChat(
+            with: service.eraseToAnyService(),
+            navigationPath: $appState.routing.teacherFinderNavigationPath
+          )
         }
+      )
     }
+  }
+
+  @ViewBuilder
+  var segmentedControlContent: some View {
+    switch tabItems[tabSelection] {
+    case .teacherOverview:
+      TeacherOverviewTabContent(teacher: service.author)
+    case .lessonOverview:
+      LessonOverviewTabContent(service: service)
+    case .curriculum:
+      CurriculumTabContent(curriculumList: service.curriculum)
+    case .review:
+      ReviewTabContent(service: service.eraseToAnyService())
+    }
+  }
 }
 
 extension LessonServiceDetailView {
-    enum TabItem {
-        case teacherOverview
-        case lessonOverview
-        case curriculum
-        case review
+  enum TabItem {
+    case teacherOverview
+    case lessonOverview
+    case curriculum
+    case review
 
-        var name: String {
-            switch self {
-            case .teacherOverview:
-                return "선생님 소개"
-            case .lessonOverview:
-                return "수업 소개"
-            case .curriculum:
-                return "커리큘럼"
-            case .review:
-                return "리뷰"
-            }
-        }
+    var name: String {
+      switch self {
+      case .teacherOverview:
+        "선생님 소개"
+      case .lessonOverview:
+        "수업 소개"
+      case .curriculum:
+        "커리큘럼"
+      case .review:
+        "리뷰"
+      }
     }
+  }
 }
 
 #if DEBUG
-#Preview {
+  #Preview {
     @Injected(\.appState) var appState
     appState.userData.loginInfo.currentUser = Student.sample0.eraseToAnyUser()
 
     return
-        NavigationStack {
-            LessonServiceDetailView(service: .sample0)
-        }
-}
+      NavigationStack {
+        LessonServiceDetailView(service: .sample0)
+      }
+  }
 #endif
