+++ Umpa/Presentation/Main/TeacherFinder/ServiceDetail/MrCreationService/MrCreationServiceDetailView.swift
@@ -5,99 +5,99 @@ import Domain
 import Factory
 import SwiftUI
 
-struct MrCreationServiceDetailView: ServiceDetailView {
-    @InjectedObject(\.appState) private var appState
+struct MrCreationServiceDetailView: View {
+  @InjectedObject(\.appState) private var appState
 
-    #if DEBUG
+  #if DEBUG
     @Injected(\.stubServiceDetailInteractor) private var serviceDetailInteractor
-    #else
+  #else
     @Injected(\.serviceDetailInteractor) private var serviceDetailInteractor
-    #endif
+  #endif
 
-    let service: MusicCreationService
+  let service: MusicCreationService
 
-    var tabItems: [TabItem] {
-        service.sampleMusics.isEmpty
-            ? [.teacherOverview, .serviceOverview, .review]
-            : [.teacherOverview, .serviceOverview, .samplePreview, .review]
-    }
+  var tabItems: [TabItem] {
+    service.sampleMusics.isEmpty
+      ? [.teacherOverview, .serviceOverview, .review]
+      : [.teacherOverview, .serviceOverview, .samplePreview, .review]
+  }
 
-    @State private var tabSelection = 0
+  @State private var tabSelection = 0
 
-    var body: some View {
-        content
-            .modifier(NavigationBackButton(.arrowBack))
-            .navigationDestination(for: ChatRoom.self) { chatRoom in
-                ChatRoomView(chatRoom: chatRoom)
-            }
-    }
+  var body: some View {
+    content
+      .modifier(NavigationBackButton(.arrowBack))
+      .navigationDestination(for: ChatRoom.self) { chatRoom in
+        ChatRoomView(chatRoom: chatRoom)
+      }
+  }
 
-    var content: some View {
-        ZStack(alignment: .bottom) {
-            ScrollView {
-                VStack(spacing: fs(0)) {
-                    Header(tabSelection: $tabSelection, service: service)
-                    segmentedControlContent
-                        .frame(maxWidth: .infinity, maxHeight: .infinity)
-                }
-                .padding(.bottom, bottomActionBarHeight)
-            }
-
-            BottomActionBar(
-                height: bottomActionBarHeight,
-                isLiked: false, // TODO: isLiked 를 받아와야 함
-                likeButtonAction: { isLiked in
-                    serviceDetailInteractor.markAsLike(isLiked, for: service.id)
-                },
-                primaryButtonAction: {
-                    serviceDetailInteractor.startChat(
-                        with: service.eraseToAnyService(),
-                        navigationPath: $appState.routing.teacherFinderNavigationPath
-                    )
-                }
-            )
+  var content: some View {
+    ZStack(alignment: .bottom) {
+      ScrollView {
+        VStack(spacing: fs(0)) {
+          Header(tabSelection: $tabSelection, service: service)
+          segmentedControlContent
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
         }
-    }
+        .padding(.bottom, ServiceDetailConstant.bottomActionBarHeight)
+      }
 
-    @ViewBuilder
-    var segmentedControlContent: some View {
-        switch tabItems[tabSelection] {
-        case .teacherOverview:
-            TeacherOverviewTabContent(teacher: service.author)
-        case .serviceOverview:
-            ServiceOverviewTabContent(service: service)
-        case .samplePreview:
-            SamplePreviewTabContent(sampleMusics: service.sampleMusics)
-        case .review:
-            ReviewTabContent(service: service.eraseToAnyService())
+      BottomActionBar(
+        height: ServiceDetailConstant.bottomActionBarHeight,
+        isLiked: false, // TODO: isLiked 를 받아와야 함
+        likeButtonAction: { isLiked in
+          serviceDetailInteractor.markAsLike(isLiked, for: service.id)
+        },
+        primaryButtonAction: {
+          serviceDetailInteractor.startChat(
+            with: service.eraseToAnyService(),
+            navigationPath: $appState.routing.teacherFinderNavigationPath
+          )
         }
+      )
     }
+  }
+
+  @ViewBuilder
+  var segmentedControlContent: some View {
+    switch tabItems[tabSelection] {
+    case .teacherOverview:
+      TeacherOverviewTabContent(teacher: service.author)
+    case .serviceOverview:
+      ServiceOverviewTabContent(service: service)
+    case .samplePreview:
+      SamplePreviewTabContent(sampleMusics: service.sampleMusics)
+    case .review:
+      ReviewTabContent(service: service.eraseToAnyService())
+    }
+  }
 }
 
 extension MrCreationServiceDetailView {
-    enum TabItem {
-        case teacherOverview
-        case serviceOverview
-        case samplePreview
-        case review
+  enum TabItem {
+    case teacherOverview
+    case serviceOverview
+    case samplePreview
+    case review
 
-        var name: String {
-            switch self {
-            case .teacherOverview:
-                return "선생님 소개"
-            case .serviceOverview:
-                return "서비스 안내"
-            case .samplePreview:
-                return "샘플 확인"
-            case .review:
-                return "리뷰"
-            }
-        }
+    var name: String {
+      switch self {
+      case .teacherOverview:
+        "선생님 소개"
+      case .serviceOverview:
+        "서비스 안내"
+      case .samplePreview:
+        "샘플 확인"
+      case .review:
+        "리뷰"
+      }
     }
+  }
 }
 
 #if DEBUG
-#Preview {
+  #Preview {
     MrCreationServiceDetailView(service: .sample0)
-}
+  }
 #endif
